Add a higly reduced chunk implementation for far away chunks which don't need the full block and texture resolution. The blocks in those far away chunks are 1, 2, 4, 8, or 16 times as big as normal blocks thus making both rendering and generation a lot faster. Instead of rendering every single block which would take a lot of CPU-time the far-away chunks are loaded into a chunk-mesh, so in the end only one rendering call is required per chunk.
The aim of all this is to have the terrain get lower in resolution the further it is from the player. This will ultimately allow for a lot higher render distance.
The lower resolution will make generating, loading and displaying of those regions a lot faster.
This ultimately will allow for multiple times the current render distance without significant performance loss.

Why put this in a new branch?
Our current engine had 0 support for this and even to make it barely work in this commit, I had to make numerous additions to the engine:
A new type of chunk, new shaders, primitive arraylists, chunk meshing.
But there is still a lot of work ahead:
- Faraway chunks currently have only one color which makes it incredible had to make out details.
- Other terrain features(caves, rivers, trees, …) need to be taken into account.
- Lighting needs to be added.
- There are issues at the world border.
- And of course bug fixes and performance optimizations.
That is all going to take a long time and I don't want to delay the next release more than it is already delayed.