Automate Shader uniform location storage using reflection.
Previously the uniform locations were stored in a hashmap and were referenced by name.
Now they are stored in static variables of a class provided upon shader creation. This allows storing the locations as variables instead of in a hashmap.
This is faster and additionally has IDE support. And thanks to reflection the annoying initialization is no longer needed.