Add options to choose climate/map generators.
The old SurfaceGenerators were removed because there were many issues with that approach.
Along the way I made some refactoring, such as moving code from local world to the ChunkManager(formerly chunkGenerationThreadPool). The ChunkManager is now also responsible for storing chunks and data.

The world generation properties are now stored in a separate json file and are used internally as json when creating the world.