Remove heightmap storing.
This was used to make light calculation more accurate, especially when building stuff.
But it is not working well.
There are constantly terrain generation changes and if the heightmap gets stored, there will be lighting bugs on each iteration.
Additionally it was quite resource hungry(it was responsible for some lag when pausing and created thousands of folders)

My long-term goal is to propagate light through LOD chunks which creates the same effects in a more dynamic way.