Remove some unecessary opengl calls from the ReducedChunkMesh:
The vao needs to be bound only once.
The voxelSize can be stored inside the SSAO instead of repeatedly setting a uniform.
There is no need to set shader parameters for the replacement mesh, when there is no replacement mesh.
The SSBO doesn't need to be set to 0 after use.