Fix performance issues caused by `lock` and `idivl` instruction.
I actually looked at the x86 assembly code for this one. And I discovered that most of the terrain generation code is bottlenecked by these two instructions.
The `lock` comes from java.util.Randoms use of atomics. To fix this I added a replacement for that(which uses the exact same algorithm). While on that I also added a static interface, to reduce useless object allocation.
The `idivl` comes from occasional divisions by `voxelSize`. Since `voxelSize` is always a power of two, this could be reaplced by a rightshift.

Now most of the generation code seems to be dominated by actual calculation(such as square roots) and cache misses(`mov`).