Interpolation in a bcc lattice using 3d barycentric coordinates.
Among that I also improved the cave biome map to actually use a bcc lattice as the underlying biome grid (before that a biome was spread over multiple points in the lattice, now it's just one.

And I also improved performance of the 2d fractal noise algorithms by using `@memcpy`.