Store the command payload as a tagged union instead of always storing all types.
This allows decoupling the payload from the undo steps, making it possible to launch subtasks with a different payloads (e.g. taking an item from creative and putting into the source slot instead of the dest slot).

Additionally this might also prevent some trouble when preparing the network data in the future.

Another small step towards #643