Misc texture palette tweaks (#1568)
Some thrilling and game-changing texture edits:
These are a bunch of small tweaks that have been on my to-do list.
Ideally i would have before/afters for these but they are so minor that
I feel its not worth the effort

- Lower contrast of brightest sand colour (this was bothering me all
week when testing logs in the desert)

![image](https://github.com/user-attachments/assets/764933e4-11db-4f68-9081-5ab9f29d7233)

- Lower contrast of duckweed 

![image](https://github.com/user-attachments/assets/73820d80-c17d-4cb4-96d4-52e784d9d2be)

- Reduce hue shifting on red chalk (the darkest colour is now more red
than more pink)

![image](https://github.com/user-attachments/assets/cf56f087-b33b-4250-8877-e56977c87905)

- Darken dead leaves & lower contrast. This is probably the most
noticeable change. (ensuring good contrast when on top of grass blocks)

![image](https://github.com/user-attachments/assets/1fe5c9b1-3f05-45a3-96bc-8550e3a49d60)