Separate the face buffer based on LOD
This reduces allocator overhead (less items in the free list of each buffer),
it reduces the chance to run out of fake GPU memory (we can now effectively store 50% more faces than before)
it reduces the chance to run out of actual GPU memory
it reduces height of lag spikes of buffer resizes (but increases their frequency)
it makes it more clear in which LOD (LOD1) the main memory bottlenecks are.
it also reduces the occlusion culling lag after disocclusion of parts of the screen.

It however does decrease the granularity of timing information, since I can no longer look at the passes separately.