The end of elasticity (ikabods simplification dreams), and the end of durability per swing (#1830)
As I was working on removing durability per swing, I noticed that the
current system makes early-game tools have way too much durability, and
with the old system it was hard to nerf this, since all properties
influence both durability and swing time or damage.
So I decided to just simply specify the parameters damage, swing speed
and durability in the material. This makes it easier to specify
reasonable values for a material.

Remaining work:
- [x] Rebalance remaining materials
- [x] Rebalance remaining tools
- [x] Figure out what to do with the tip (gold-tipping is too OP, almost
as good as uranium)

fixes #1661
closes #1559 (proposal related to durability per swing)
closes #1567 (proposal related to durability per swing)
closes #1430 (new rebalance → old balancing problems no longer apply)
provides an alternative for #1339 to make wooden handles more viable
progress towards #1584