Add headless server mode
Adds headlessServerMode boolean to launchConfig
Adds headlessServerModeWorldName string to lauchConfig (The name of the world to create/use in headless mode)
Adds headlessGameMode to launchConfig (creative or survival)
Adds headlessAllowCheats boolean to launchConfig
Adds headlessTestingMode boolean to launchConfig (Unsure what this is for, but, flawedCreateWorld used it, so I pass it through.)

Splits flawedCreateWorld into its own file under server for more general use outside of the gui (args are worldName, gamemode, allowCheats, and testingMode)
Prevents running OpenGL code if running in headless mode.
Sets player name to "serverDummy" - but no client joins the server. The server's just running and accepting connections.

Not a full fix for #1097, but a workaround.