Fix server asset streaming.
This solves two issues:

1. The server/client creating the world never loads assets into the world to be packed later, which leads to an empty archive sent to connecting clients for serverAssets
2. The path to access serverAssets on the client end was malformed. For instance, if cubyzDir was ./data, it would try to access ./data/./data/serverAssets, instead of ./data/serverAssets.

I briefly tried to move assets to cubyzDir/assets but there are a lot of things tangled in that which aren't necessary to fix this bug. There are a lot of examples of code I've found which independently constructs an assets folder path rather than using a common one via something like main.game.assetsPath.