Fix menu fov overriding fov setting (#1976)
Fixes #1962 by adding a defer which sets the viewport fov to whatever
the selected fov is in settings. This line is located right under the
one which sets the fov to 70 while in the main menu. So upon entering a
world from now on, the fov should properly be adjusted to whatever you
chose.

I'm still pretty new to GitHub and this is my first pull request, so
please let me know if there are any issues or if I can do anything
better.