Split Gamemode and DamageType out of game.zig
The purpose of doing this is to prepare for a future headless server.
We would not want to use game.zig as it currently stands in a headless
server. We only really care about the Gamemode and DamageType types it
provides.

As a stopgap, and to reduce the blast radius of this change, I've made
them accessible via game.Gamemode and game.DamageType as they had been
before.

In the future it might be a good idea to encapsulate it in its own
struct (i.e. Game) and game.zig becomes client.zig or something, maybe.