Implementing culling in VerticalList.render. Fixes #1991
Items that do not fall within the bounds of the guiwindow are not
rendered.

Whether a item is within view or not is determined by calculating the
item's position in the list and the item's position offset by the scrollbar's
position

If that offset value is < 0, skip rendering
If that offset value is > self.maxHeight, skip rendering.