Fix empty crafting window being too large (#2021)
Fixes #2020.

I opted to go for the simplest fix for now: Simply setting up the window
with an empty list of crafting recipes.

Perhaps a more ideal solution would be for there to be a text saying "no
recipes found" or so.
Let me know if I should address this with this PR or if that's wasted
effort.