fix issue with verticle list relative positions (#2199)
While implement culling in VerticalList.render #2024 the culling was
only tested on the chat from what I know.

While making a example for #2065 I noticed a bug while showing the list.
Problem was `shiftedPos` is relative to the parent component, different
than all other variables used in the if.
I did not create an issue because I knew the problematic code and was
able to make a quick fix. I can also make an issue if wanted.