Fill partial stacks before placing in an empty slot (#2190)
Fixes #2120

The approach is to first loop the inventory only for slots with at least
some of source item to attempt filling those slots first.

~~Then, if remainingAmount > 0, do the same loop again but to find an
empty slot to place the remainder into.~~
EDIT:
I found a more efficient way. During the existing loop, find the first
empty slot and mark that as the `selectedEmptySlot`.

Then, distribute the blocks into all slots with some of the item in it.

Then, after the loop, place the remaining amount (if any) in the
`selectedEmptySlot` from earlier.