Make the Seed Configurable (#2136)
This PR adds an input field to the world creation GUI that allows the
user to input a seed. If the user leaves this field blank, then the seed
is randomly generated. If the user enters a string that cannot be parsed
to a `u64`, it will be hashed.
- - -
<img width="455" height="768" alt="Screenshot_20251025_230211"
src="https://github.com/user-attachments/assets/02cddcd5-d1fc-4e70-85f6-91313c705e16"
/>