fix projection matrix not updating on window resize (#2321)
projection matrix was not updated on window resize leading to stretched
rendering in the main menu and the game world. example:

this: 


https://github.com/user-attachments/assets/2ded6189-da88-44d3-a616-e39b201168d3

is now fixed as:


https://github.com/user-attachments/assets/5e163657-4f47-4db6-8dae-62f9461b7bae

including the menu and all.
i checked logically and via debug prints that it only updates if the
window resizes or user updates the fov in the graphic settings