Branch decay [Simple Implementation] (#2334)
Introduces a two tiered system of decay.
- Leaves depend on either logs or branches.
- And branches depend on logs.

This means that breaking the logs of a tree is already enough for the
whole tree to decay.
This would fix
https://github.com/PixelGuys/Cubyz/issues/2331#issuecomment-3566930054

Downsides:
The log search algorithm is more expensive now, since it compares tags
instead of the decayProhibitor attribute

Things i also had to introduce:
- Callbacks have a deinit function now.
- the 7th bit of the cubyz:branch data is used for storing the
was-this-placed-by-human boolean

---------

Co-authored-by: IntegratedQuantum <43880493+IntegratedQuantum@users.noreply.github.com>