Add obfuscation option for the TextInput component and use it to implement a streamer mode option to hide your IP addresses (#2029)
Adds the ability to hide the input text. Usage via the `.obfuscate()`
and `.deobfuscate()` methods. To ensure uniform and consistent use of
the same obfuscation character across other parts of the code, the
global members `TextInput.obfuscationChar` and
`TextInput.obfuscatedStringBuffer` are public.

This will be very useful for #1929. The implementation proposed in #1933
is broken, does not support Utf8, and constantly uses memory allocation
to create a string of filler characters, which is precisely why I am
proposing this implementation.

fixes #1929