Give me that sweet sweet abstraction (#2417)
So I was tired of those long function signatures in of
`Inventory.Command.Payload.run` so I squeezed them into a `Context`
struct.
While on it I also added a `Context.execute` method wrapping
`executeBaseOperation` since that ended up being a lot of characters
too.
This way it should be easier to change the code in case of refactors.