Send chat commands through the inventory synchronization system (#2415)
with this all actions that affect inventory and other synchronized
properties (→gamemode) are ran on the server thread, which would allow
us to remove the mutex and add some replacement checks to protect
against misuse.

progress towards #1744 and #2414

it may seem odd to handle commands like this, but I think ultimately we
should synchronize commands. without this you could e.g. take an item
from the creative inventory, send the command, and then get the item
reverted because the command may have been executed first on the server.
And of course it greatly simplifies the immediate task #1744 here.