Use i128 in zon and fix random number generation for large bitsizes
also I decided to avoid casting, which allows us to easier find these issues (in case we ever need u128) in the future.

fixes #2290
the odd thing about #2290 was that it only applied to seeds generated from text, but not randomly generated seeds. When investigating this I found a severe issue in our random number generator which is also fixed here.