Remove the inventory mutex and replace it with equivalent assertions (#2427)
this required some more smaller restructuring to make the assertions
pass.
And we definitely need to test this in another play session before the
release.

But this finally fixes #1744 since everything of importance is now ran
on the main server thread (except from inventory creation, which was
easily synchronized with a lock)

In the future (in the context of #2420) the threadContext is intended to
hold the position of the chunk region that is accessible, this would
allow adding additional assertions to check if e.g. positional
inventories are usable or if a block is allowed to be placed in a given
chunk.